+++ src/main/java/fr/alasdiablo/janoeo/config/GravelConfig.java
@@ -22,6 +22,8 @@ public class GravelConfig {
         public final ForgeConfigSpec.BooleanValue GOLD_GRAVEL_ORE;
         public final ForgeConfigSpec.BooleanValue DIAMOND_GRAVEL_ORE;
         public final ForgeConfigSpec.BooleanValue IRON_GRAVEL_ORE;
+        public final ForgeConfigSpec.BooleanValue COAL_GRAVEL_ORE;
+        public final ForgeConfigSpec.BooleanValue LAPIS_GRAVEL_ORE;
 
 
         public Config(ForgeConfigSpec.Builder builder) {
@@ -30,6 +32,8 @@ public class GravelConfig {
             GOLD_GRAVEL_ORE = builder.comment("Gravel gold ore generation: disable / enable").define("gravelGoldOreGen", true);
             DIAMOND_GRAVEL_ORE = builder.comment("Gravel diamond ore generation: disable / enable").define("gravelDiamondOreGen", true);
             IRON_GRAVEL_ORE = builder.comment("Gravel iron ore generation: disable / enable").define("gravelIronOreGen", true);
+            COAL_GRAVEL_ORE = builder.comment("Gravel coal ore generation: disable / enable").define("gravelCoalOreGen", true);
+            LAPIS_GRAVEL_ORE = builder.comment("Gravel lapis ore generation: disable / enable").define("gravelLapisOreGen", true);
 
             builder.pop();
         }