+++ src/main/java/fr/alasdiablo/janoeo/init/GravelsOresBlocks.java
@@ -37,6 +37,7 @@ public class GravelsOresBlocks {
 
     public static Block COAL_GRAVEL_ORE = new GravelOre(Registries.COAL_GRAVEL_ORE, COMMON);
     public static Block LAPIS_GRAVEL_ORE = new GravelOre(Registries.LAPIS_GRAVEL_ORE, UNCOMMON);
+    public static Block EMERALD_GRAVEL_ORE = new GravelOre(Registries.EMERALD_GRAVEL_ORE, RARE);
 
     /**
      * Registry Event Handler
@@ -52,7 +53,7 @@ public class GravelsOresBlocks {
         public static void onBlocksRegistry(final RegistryEvent.Register<Block> event) {
             RegistryHelper.registerBlock(event.getRegistry(),
                     DIAMOND_GRAVEL_ORE, IRON_GRAVEL_ORE, GOLD_GRAVEL_ORE,
-                    COAL_GRAVEL_ORE, LAPIS_GRAVEL_ORE
+                    COAL_GRAVEL_ORE, LAPIS_GRAVEL_ORE, EMERALD_GRAVEL_ORE
             );
         }
 
@@ -66,7 +67,7 @@ public class GravelsOresBlocks {
             Item.Properties properties = new Item.Properties().group(JanoeoGroup.ORE_BLOCKS);
             RegistryHelper.registerBlockItem(event.getRegistry(), properties,
                     DIAMOND_GRAVEL_ORE, IRON_GRAVEL_ORE, GOLD_GRAVEL_ORE,
-                    COAL_GRAVEL_ORE, LAPIS_GRAVEL_ORE
+                    COAL_GRAVEL_ORE, LAPIS_GRAVEL_ORE, EMERALD_GRAVEL_ORE
             );
         }
     }