+++ changelog.md
@@ -1,10 +1,28 @@
-### %WIP Version%
-+ Fix missing ore generation ([#10](https://github.com/AlasDiablo/JANOEO/issues/10))
-  + Diamond nether ore
-  + Emerald nether ore
-+ Add dust
-  + Coal dust
-  + Lapis dust
+### 5.6.0.67
++ Fix
+  + Missing ore generation ([#10](https://github.com/AlasDiablo/JANOEO/issues/10))
+    + Diamond nether ore
+    + Emerald nether ore
++ Add
+  + Dust
+    + Coal
+    + Lapis
+    + Emerald
+    + Aluminium
+    + Lead
+    + Nickel
+    + Platinum
+    + Ruby
+    + Sapphire
+    + Silver
+    + Tin
+    + Uranium
+    + Copper
+  + Gravel Ore
+    + Coal
+    + Lapis
+    + Emerald
+  + Custom creative tab
 
 ### 4.5.0.63-experimental
 + Add support for DiaboloLib 1.0.4